pyproject.toml
[tool.poetry]
name = "rasterly"
version = "0.1.0"
description = ""
authors = ["(redacted)"]

[tool.poetry.dependencies]
python = "^3.7"
mouse = "^0.7.1"
pillow = "^8.0"
numpy = "^1.19"
click = "^7.1"
ipykernel = "^5.3"
matplotlib = "^3.3"
PyUserInput = "^0.1.11"
autopep8 = "^1.5"
ipdb = "^0.13.4"
pretty-errors = "^1.2"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5"
black = {version = "^18.3-alpha.0", allows-prereleases = true}
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
